feat: some refactoring and warning fix

+++ Engine/src/Math/Math.hpp
@@ -58,12 +58,12 @@ namespace nb
 
         inline float toRadians(const float degrees) noexcept
         {
-            return degrees * (Constants::PI / Constants::HALF_OF_CIRCLE_IN_DEG);
+            return static_cast<float>(degrees * (Constants::PI / Constants::HALF_OF_CIRCLE_IN_DEG));
         } 
 
         inline float toDegrees(const float radians) noexcept
         {
-            return radians * (Constants::HALF_OF_CIRCLE_IN_DEG / Constants::PI);
+            return static_cast<float>(radians * (Constants::HALF_OF_CIRCLE_IN_DEG / Constants::PI));
         }
 
         inline Vector3<float> toFloatColor(const Vector3<uint8_t>& color)
